// apps/ui/src/applications/database-client/lib/types.ts
export interface DbColumn {
  name: string;
  typeName: string;
}